chore: 添加动态气泡图 demo
