+++ examples/gallery/dynamic/index.en.md
@@ -1,0 +1,4 @@
+---
+title: Dynamic Charts
+order: 7
+---
