+++ examples/gallery/dynamic/index.zh.md
@@ -1,0 +1,4 @@
+---
+title: 动态图表
+order: 7
+---
